Pajek(blockmodeling)	R Documentation

Functions for loading and writing Pajek[http://vlado.fmf.uni-lj.si/pub/networks/pajek/] files

Description

Functions for reading/loading and writing Pajek files:

loadnetwork - Loads a Pajek ".net" file as a matrix. For now, only simple one and two-mode networks are supported (eg. only single relations, no time information).

loadnetwork2 - The same as above, but adopted to be called withih loadpajek

savenetwork - Saves a matrix in to a Pajek ".net" file.

loadmatrix - Loads a Pajek ".mat" file as a matrix.

savematrix - Saves a matrix in to a Pajek ".mat" file.

loadvector - Loads a Pajek ".clu" file as a vector.

loadvector2 - The same as above, but adopted to be called withih loadpajek - as a consequence not suited for reading clusters

savevector - Saves a vector in to a Pajek ".clu" file.

loadpajek - Loads a Pajek project file (".paj") as a list with the following components: Networks, Partitions, Vectors and Clusters. Clusters and hierarchies are dissmised.

Usage


loadnetwork(file,useSparseMatrix=NULL,minN=50)
loadnetwork2(file,useSparseMatrix=NULL,minN=50)
savenetwork(n, file, twomode = "default", symetric = all(n==t(n)))
loadmatrix(file)
savematrix(n, file, twomode = 1)
loadvector(file)
loadvector2(file)
savevector(v, file)
loadpajek(file)


Arguments

file 	 The name of the file to be loaded or saved to or an open file object.

useSparseMatrix 	 Should a sparse matrix be use instead of the ordinary one? Sparse matices can only be used if package Matrix is installed. The default NULL uses sparsematrices for networks with more that minN vertices

minN 	 The minimal number of units in the network to use sparse matrices.

n 	 A matrix representing the network.

twomode 	 1 for one-mode networks and 2 for two-mode networks. Default sets the argument to 1 for square matrices and to 2 for others.

symetric 	 If true, only the lower part of the matrix is used and the values are interpreted as "Edges", not "Arcs".

v 	 A vector

Value

NULL, a matrix or a vector (see Description)

Author(s)

Vladimir Batagelj & Andrej Mrvar (most functions), Aleš Žiberna (loadnetwork,loadpajek and modification of others)

References

Pajek ( V. Batagelj, A. Mrvar: Pajek - Program for Large Network Analysis. Home page http://vlado.fmf.uni-lj.si/pub/networks/pajek/.

W. de Nooy, A. Mrvar, V. Batagelj: Exploratory Social Network Analysis with Pajek, CUP, January 2005

See Also

plot.mat,crit.fun,opt.par,opt.random.par,opt.these.par,check.these.par

[Package blockmodeling version 0.1.6 Index]