find.m(blockmodeling)	R Documentation

Computing the threshold

Description

The functions compute the maxsimum value of m/cut where a ceratin block is still classified as alt.blocks and not "null". The difference between find.m and find.m2 it that find.m uses an optimizational approach and is faster and more precise than find.m2. However, find.m only supports regular ("reg") and complete ("com") as alt.blocks, while find.m2 supports all block types. Also, find.m does not always work, sepecially if cormat is not "none".

Usage


find.m(M, clu, alt.blocks = "reg", diag = !is.list(clu), cormet = "none", FUN = "max", half = TRUE)
find.m2(M, clu, alt.blocks = "reg", neval = 100, half = TRUE, ...)
find.cut(M, clu, alt.blocks = "reg", cuts = "all", ...)

Arguments

M 	 A matrix representing the (usually valued) network. For now, only one-relational networks are supported. The network can have one or more modes (diferent kinds of units with no ties among themselvs. If the network is not two-mode, the matrix must be square.

clu 	 A partition. Each unique value represents one cluster. If the nework is one-mode, than this should be a vector, else a list of vectors, one for each mode

alt.blocks 	 Only one of allowed blocktypes, as alternative to the null block:
"com" - complete block
"rdo", "cdo" - row and column-dominant blocks (binary, valued, and implicit approach only)
"reg" - (f-)regular block
"rre", "cre" - row and column-(f-)regular blocks
"rfn", "cfn" - row and column-dominant blocks (binary, valued, and implicit approach only)
"den" - density block (binary approach only)
"avg" - average block (valued approach only)

diag 	 (default = TRUE) Should the special stauts of diagonal be acknowladged.

cormet 	 Which metho should be used to correct for diferent maxismum error contributins?
"none" - no correction
"censor" - censor values larger than m
"correct" - so that the maxsimum possible error contribution of the cell is the same regardles of a condition (either that somthing must be o or at least m)

FUN 	 (default = "max") Function f used in row-f-regular, column-f-regular, and f-regular blocks.

cuts 	 The cuts which should be evaluatated. If cuts="all"n (default), all unique values are evaluated

neval 	 Number of different m values to be evaluated.

half 	 Should the returned value of m be one half of the value where the incosnistencies are the same.

... 	 Other parameters to crit.fun

Value

A matrix of maximal m/cut values.

Author(s)

Aleš Žiberna

References

ŽIBERNA, Aleš (2006): Generalized Blockmodeling of Valued Networks. Social Networks, Jan. 2007, vol. 29, no. 1, 105-126. http://dx.doi.org/10.1016/j.socnet.2006.04.002.

ŽIBERNA, Aleš. Direct and indirect approaches to blockmodeling of valued networks in terms of regular equivalence. J. math. sociol., 2008, vol. 32, no. 1, 57-84. http://www.informaworld.com/smpp/content?content=10.1080/00222500701790207.

DOREIAN, Patrick, BATAGELJ, Vladimir, FERLIGOJ, Anuška (2005): Generalized blockmodeling, (Structural analysis in the social sciences, 25). Cambridge [etc.]: Cambridge University Press, 2005. XV, 384 p., ISBN 0-521-84085-6.

See Also

crit.fun and maybe also opt.par,opt.random.par,opt.these.par,check.these.par,plot.mat

[Package blockmodeling version 0.1.5 Index]