genRandomPar(blockmodeling)	R Documentation

The function for generating random partitions

Description

The function generates random partitions. The function is meant to be caled by the function opt.random.par

Usage


genRandomPar(k, n, seed = NULL, mingr = 1, maxgr = Inf, addParam = list(genPajekPar = TRUE, probGenMech = NULL))

Arguments

k 	 Number of clusters (by modes)

n 	 Number of units (by modes)

seed 	 Seed for generating random numbers (partitions)

mingr 	 Minimal alowed group size

maxgr 	 Maximal alowed group size

addParam 	 This has to be a list with the following parameters (any or all can be missing, then the default values (see usage) are used):
"genPajekPar" - Should the partitions be generated as in Pajek (Batagelj and Mrvar, 2006). If FALSE, all partitions are selected completly at random while making sure that the partitions have the required number of clusters.
"probGenMech" - Here the probabilities for 4 diferent generating mechanichems can be specified. If this is not specified, the value is set to c(1/3,1/3,1/3,0) if genPajekPar is TRUE and to c(0,0,0,1) if if genPajekPar is FALSE. The first 3 mechanisems are the same as implemetned in Pajek (the second one has almost all units in only one cluster) and the fourth is completly random (from uniform distribution).

Value

A random partition in the format required by opt.random.par. If a netowork has several modes, than a list of partitions, one for each mode.

Author(s)

Aleš Žiberna

References

BATAGELJ, Vladimir, MRVAR, Andrej (2006): Pajek 1.11, http://vlado.fmf.uni lj.si/pub/networks/pajek/ (accessed January 6, 2006).

See Also

opt.random.par

[Package blockmodeling version 0.1.6 Index]