gplot1(blockmodeling)	R Documentation

A wrapper for funcction gplot - Two-Dimensional Visualization of Graphs

Description

The function calls function gplot from library sna with different defaults. Usefun for ploting image graphs.

Usage


gplot1(M, uselen = TRUE, usecurve = T, edge.len = 0.001, diag = T, displaylabels = T, boxed.labels = F, loop.cex = 4, arrowhead.cex = 2.5, edge.lwd = 1, edge.col = gray(1 - M/max(M)), thresh = 0.05, ...)
gplot2(M, uselen = TRUE, usecurve = T, edge.len = 0.001, diag = T, displaylabels = T, boxed.labels = F, loop.cex = 4, arrowhead.cex = 2.5, edge.lwd = 1, edge.col = "default", thresh = 0.05, ...)

Arguments

M 	 A matix (array) of a graph or set thereof. This data may be valued.

diag 	 boolean indicating whether or not the diagonal should be treated as valid data. Set this true if and only if the data can contain loops. diag is FALSE by default.

thresh 	 real number indicating the lower threshold for tie values. Only ties of value >thresh are displayed. By default, thresh=0.

mode 	 the vertex placement algorithm; this must correspond to a gplot.layout function.

displaylabels 	 boolean; should vertex labels be displayed?

boxed.labels 	 boolean; place vertex labels within boxes?

arrowhead.cex 	 expansion factor for edge arrowheads.

loop.cex 	 expansion factor for loops; may be given as a vector, if loops are to be of different sizes.

edge.col 	 color for edges; may be given as a vector or adjacency matrix, if edges are to be of different colors.

edge.lwd 	 line width scale for edges; if set greater than 0, edge widths are scaled by edge.lwd*dat. May be given as a vector or adjacency matrix, if edges are to have different line widths.

edge.len 	 if uselen==TRUE, curved edge lengths are scaled by edge.len.

uselen 	 boolean; should we use edge.len to rescale edge lengths?

usecurve 	 boolean; should we use edge.curve?

... 	 additional arguments to plot.

Value

Plots a graph

Author(s)

Aleš Žiberna

See Also

sna:gplot

[Package blockmodeling version 0.1.4 Index]