rand(blockmodeling)	R Documentation

Comparing partitions

Description

Rand Index and Rand Index corrected/adjusted for chance for comparing partitions (Hubert and Arabie, 1985). The names of the clusters do not matter.

Usage


rand(tab)
rand2(clu1, clu2)
crand(tab)
crand2(clu1, clu2)

Arguments

clu1, clu2 	 The two partitions to be compared, given in the from of vectors, where for each unit a cluster membership is given.

tab 	 A contingency table obtaind as table(clu1,clu2)

Value

The value of Rand Index (corrected/adjusted for chance)

Author(s)

Aleš Žiberna

References

Hubert L. in Arabie P. (1985): Comparing Partitions. Journal of Classification, 2, 193-218.

[Package blockmodeling version 0.1.2 Index]