reorderImage(blockmodeling)	R Documentation

Reorders an image matrix of the blockmodel (or an error matrix based on new and old partition.

Description

Reorders an image matrix of the blockmodel (or an error matrix based on new and old partition. The partitions should be the same, except that classes can have different labels. It is useful when we want to have a different oreder of classes in figures and then also in image matrices. Currently it is only suitable for one-mode blockmodels

Usage


reorderImage(IM, oldClu, newClu)

Arguments

IM 	 An image or error matrix.

oldClu 	 Old partition

newClu 	 New partition, the same as the old one except for class labeles.

Value

Reorder matrix (rows and columns are reordred)

Author(s)

Ales Ziberna

References

ŽIBERNA, Aleš (2006): Generalized Blockmodeling of Valued Networks. Social Networks, Jan. 2007, vol. 29, no. 1, 105-126. http://dx.doi.org/10.1016/j.socnet.2006.04.002.

ŽIBERNA, Aleš. Direct and indirect approaches to blockmodeling of valued networks in terms of regular equivalence. J. math. sociol., 2008, vol. 32, no. 1, 57-84. http://www.informaworld.com/smpp/content?content=10.1080/00222500701790207.

See Also

crit.fun,plot.mat,clu,IM,err

[Package blockmodeling version 0.1.5 Index]