ss(blockmodeling)	R Documentation

Sum of Squared deviations from the mean and sum of Absolute Deviations from the median

Usage


ss(x)
as(x)

Value

Sum of Squared deviations from the mean or sum of Absolute Deviations from the median

Author(s)

Aleš Žiberna

[Package blockmodeling version 0.1.5 Index]